Add automatic edge archiving when a connected goal is archived
In making this change, I had to be able to look up edges by their actual 
entry address within the UI. This means I mirrored the design of the 
goals list within redux and applied it to the edges. Edges are now 
represented in Redux as an object; keys are addresses and values are 
edge objects with parent, child, and self address.

This also meant I had to change nearly all of the code that touched 
edges to work with this new format. Some refactoring might be in order 
there.